Code golf
This trims down the helper from 236 characters to 207.

There are two safe transforms:
- Cache `Promise` repeated twice into a variable
- Remove `{}` from a `if` and use one `;` instead

And an unsafe transform that uses arrow function instead of normal functions. This assumes that you are running in an enviromnent where there's both generators and arrow functions.

I mostly made this pull request for fun, feel free to close it without reviewing it :)